Aggiunta immagine di default

+++ Controller/CUtente.php
@@ -44,14 +44,19 @@ class CUtente {
 	
 	public function getImmagine(){
 		$FImmagine=USingleton::getInstance('FImmagine');
-		$image = $FImmagine->getImmagineByNome($_REQUEST['file']);
-		$handle = fopen("./tmp/".$_REQUEST['file'],"w+");
-		fwrite($handle,$image[0]['immagine_originale']);
-		$file = "./tmp/".$_REQUEST['file'];
-		header('Content-Type: image/'.basename($image[0]['type']));
-		header('Content-Length: ' . $image[0]['size']);
-		echo file_get_contents($file);
-		unlink($file);
+		if ($_REQUEST['file'] == NULL) {
+			$file = "./tmp/".$_REQUEST['file'];
+			echo file_get_contents($file);
+		} else {
+			$image = $FImmagine->getImmagineByNome($_REQUEST['file']);
+			$handle = fopen("./tmp/".$_REQUEST['file'],"w+");
+			fwrite($handle,$image[0]['immagine_originale']);
+			$file = "./tmp/".$_REQUEST['file'];
+			header('Content-Type: image/'.basename($image[0]['type']));
+			header('Content-Length: ' . $image[0]['size']);
+			echo file_get_contents($file);
+			unlink($file);
+		}
 	}
 	
 	public function inviaInfo() {
